% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comments.R
\name{add_comments}
\alias{add_comments}
\title{Add comments on Files and URLs}
\usage{
add_comments(hash = NULL, comment = NULL, ...)
}
\arguments{
\item{hash}{hash for the resource you want to comment on; Required; String}

\item{comment}{review; Required; String}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
data.frame with 2 columns: response_code, verbose_msg
}
\description{
Add comments on files and URLs. For instance, flagging false positives, adding details about malware, instructions for cleaning malware, etc.
}
\examples{
\dontrun{
add_comments(hash='99017f6eebbac24f351415dd410d522d', comment="This is great.")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

