% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{as.data.frame.virustotal_response}
\alias{as.data.frame.virustotal_response}
\title{Convert VirusTotal response to data.frame}
\usage{
\method{as.data.frame}{virustotal_response}(x, ...)
}
\arguments{
\item{x}{A virustotal_response object}

\item{...}{Additional arguments (unused)}
}
\value{
A data.frame representation of the response
}
\description{
Convert VirusTotal response to data.frame
}
