% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain_relationship.R
\name{domain_relationship}
\alias{domain_relationship}
\title{Retrieve related objects to an Internet domain}
\usage{
domain_relationship(
  domain = NULL,
  relationship = "subdomains",
  limit = NULL,
  cursor = NULL,
  ...
)
}
\arguments{
\item{domain}{domain name. String. Required.}

\item{relationship}{domain name. String. Required. Default is \code{subdomains}. 
For all the options see \link{https://developers.virustotal.com/v3.0/reference#domains-relationships}}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
named list
}
\description{
Retrieve related objects to an Internet domain
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
domain_relationship("http://www.google.com")
domain_relationship("http://www.goodsfwrfw.com") # Domain not found
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
