% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_report.R
\name{domain_report}
\alias{domain_report}
\title{Get Domain Report}
\usage{
domain_report(domain = NULL)
}
\arguments{
\item{domain}{domain name (string)}
}
\value{
data frame
}
\description{
Get Domain Report
}
\examples{
\dontrun{
domain_report(domain="http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

