% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_behaviour_report.R
\name{get_behaviour_report}
\alias{get_behaviour_report}
\title{Get Behaviour Report}
\usage{
get_behaviour_report(sandbox_id = NULL, ...)
}
\arguments{
\item{sandbox_id}{Sandbox report ID (character string). Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing detailed behaviour report from the sandbox
}
\description{
Retrieves a specific behaviour report from a sandbox analysis.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_behaviour_report(sandbox_id = 'hash_sandboxname')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key,
\code{\link{get_file_behaviours}} for listing all behaviour reports
}
