% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_upload_url.R
\name{get_file_upload_url}
\alias{get_file_upload_url}
\title{Get file upload URL for large files}
\usage{
get_file_upload_url(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing upload URL and other metadata
}
\description{
Get a special URL for uploading files larger than 32MB to VirusTotal for analysis.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_file_upload_url()
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key,
\code{\link{scan_file}} for regular file uploads
}
