% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip_comments.R
\name{get_ip_comments}
\alias{get_ip_comments}
\title{Retrieve comments for an IP address}
\usage{
get_ip_comments(ip = NULL, limit = NULL, cursor = NULL, ...)
}
\arguments{
\item{ip}{IP Address. String. Required.}

\item{limit}{Number of entries. Integer. Optional.  Default is 10.}

\item{cursor}{String. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
named list
}
\description{
Retrieve comments for an IP address
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_ip_comments("64.233.160.0")
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
