% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip_relationship.R
\name{get_ip_relationship}
\alias{get_ip_relationship}
\title{Retrieve related objects to an IP Address}
\usage{
get_ip_relationship(
  ip = NULL,
  relationship = "subdomains",
  limit = NULL,
  cursor = NULL,
  ...
)
}
\arguments{
\item{ip}{IP address. String. Required.}

\item{relationship}{domain name. String. Required. Default is \code{subdomains}. 
For all the options see \link{https://developers.virustotal.com/v3.0/reference#domains-relationships}}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
named list
}
\description{
Retrieve related objects to an IP Address
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
get_ip_relationship("64.233.160.0")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
