% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limiting.R
\name{get_rate_limit_status}
\alias{get_rate_limit_status}
\title{Get current rate limit status}
\usage{
get_rate_limit_status()
}
\value{
List with current status information
}
\description{
Get current rate limit status
}
\seealso{
Other rate limiting:
\code{\link{rate-limiting}},
\code{\link[=rate_limit]{rate_limit()}},
\code{\link[=reset_rate_limit]{reset_rate_limit()}}
}
\concept{rate limiting}
\keyword{internal}
