% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_comments.R
\name{get_url_comments}
\alias{get_url_comments}
\title{Retrieve comments for a URL}
\usage{
get_url_comments(url_id = NULL, limit = NULL, cursor = NULL, ...)
}
\arguments{
\item{url_id}{URL or URL ID from VirusTotal}

\item{limit}{Number of comments to retrieve. Optional.}

\item{cursor}{String for pagination. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing URL comments
}
\description{
Retrieve comments for a URL
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_url_comments(url_id = 'http://www.google.com')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
