% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_key.R
\name{has_vt_key}
\alias{has_vt_key}
\title{Is an API key set in either environment variable?}
\usage{
has_vt_key()
}
\value{
Logical.
}
\description{
Is an API key set in either environment variable?
}
\seealso{
Other authentication:
\code{\link[=set_key]{set_key()}},
\code{\link[=vt_key]{vt_key()}}
}
\concept{authentication}
\keyword{internal}
