% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limiting.R
\name{is_rate_limit_initialized}
\alias{is_rate_limit_initialized}
\title{Is the request log initialized?}
\usage{
is_rate_limit_initialized()
}
\description{
Is the request log initialized?
}
\keyword{internal}
