% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_url_comments.R
\name{post_url_comments}
\alias{post_url_comments}
\title{Add a comment to a URL}
\usage{
post_url_comments(url_id = NULL, comment = NULL, ...)
}
\arguments{
\item{url_id}{URL or URL ID from VirusTotal}

\item{comment}{Comment text to add}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing response data
}
\description{
Add a comment to a URL
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

post_url_comments(url_id='http://www.google.com',
                  comment='This URL appears suspicious')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
