% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_url_votes.R
\name{post_url_votes}
\alias{post_url_votes}
\title{Add a vote to a URL}
\usage{
post_url_votes(url_id = NULL, verdict = NULL, ...)
}
\arguments{
\item{url_id}{URL or URL ID from VirusTotal}

\item{verdict}{Vote verdict: "harmless" or "malicious"}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing response data
}
\description{
Add a vote to a URL
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

post_url_votes(url_id = 'http://www.google.com', verdict = 'harmless')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
