% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescan_domain.R
\name{rescan_domain}
\alias{rescan_domain}
\title{Request rescan of a domain}
\usage{
rescan_domain(domain = NULL, ...)
}
\arguments{
\item{domain}{Domain name to rescan. Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing analysis details and ID
}
\description{
Request a new analysis of a domain already in VirusTotal's database.
Returns an analysis ID for use with \code{\link{domain_report}}.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

# Request rescan of a domain
rescan_domain("google.com")
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key,
\code{\link{domain_report}} for getting reports
}
