% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescan_url.R
\name{rescan_url}
\alias{rescan_url}
\title{Request rescan of a URL}
\usage{
rescan_url(url_id = NULL, ...)
}
\arguments{
\item{url_id}{URL or URL ID (base64 encoded URL without padding). Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing analysis details and ID
}
\description{
Request a new analysis of a URL already present in VirusTotal's database.
Returns an analysis ID for use with \code{\link{url_report}}.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

# Request rescan using URL
rescan_url("http://www.google.com")

# Request rescan using URL ID
rescan_url("687474703a2f2f7777772e676f6f676c652e636f6d2f")
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key,
\code{\link{url_report}} for getting reports
}
