% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_url.R
\name{scan_url}
\alias{scan_url}
\title{Submit URL for scanning}
\usage{
scan_url(url = NULL, ...)
}
\arguments{
\item{url}{URL to scan; string; required}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing analysis details and ID
}
\description{
Submit a URL for analysis. Returns analysis details including an ID that
can be used to retrieve the report using \code{\link{url_report}}
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

scan_url("http://www.google.com")
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
