% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key.R
\name{set_token}
\alias{set_token}
\title{Set Access Token}
\usage{
set_token(api_key = NULL)
}
\description{
Set the access token.
}
\examples{
\dontrun{
set_token('api_key_here')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

