% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\name{virustotal2_POST}
\alias{virustotal2_POST}
\title{POST for V2 API}
\usage{
virustotal2_POST(
  query = list(),
  path = path,
  body = NULL,
  key = Sys.getenv("VirustotalToken"),
  ...
)
}
\arguments{
\item{query}{query list}

\item{path}{path to the specific API service url}

\item{body}{file}

\item{key}{A character string containing Virustotal API Key. The default is retrieved from \code{Sys.getenv("VirustotalToken")}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
list
}
\description{
POST for V2 API
}
