% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{virustotal_error}
\alias{virustotal_error}
\title{Create a VirusTotal API error}
\usage{
virustotal_error(
  message,
  status_code = NULL,
  response = NULL,
  call = sys.call(-1)
)
}
\arguments{
\item{message}{Error message}

\item{status_code}{HTTP status code}

\item{response}{Full HTTP response object}

\item{call}{The calling function (automatically detected)}
}
\value{
An error object of class \code{virustotal_error}
}
\description{
Create a VirusTotal API error
}
\seealso{
Other error handling:
\code{\link{virustotal-errors}},
\code{\link[=virustotal_auth_error]{virustotal_auth_error()}},
\code{\link[=virustotal_check]{virustotal_check()}},
\code{\link[=virustotal_rate_limit_error]{virustotal_rate_limit_error()}},
\code{\link[=virustotal_validation_error]{virustotal_validation_error()}}
}
\concept{error handling}
\keyword{internal}
