% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{virustotal_file_scan}
\alias{virustotal_file_scan}
\title{Create a VirusTotal file scan result}
\usage{
virustotal_file_scan(data)
}
\arguments{
\item{data}{Raw API response data}
}
\value{
Object of class \code{virustotal_file_scan}
}
\description{
Create a VirusTotal file scan result
}
\seealso{
Other response classes:
\code{\link{virustotal-classes}},
\code{\link[=virustotal_domain_report]{virustotal_domain_report()}},
\code{\link[=virustotal_file_report]{virustotal_file_report()}},
\code{\link[=virustotal_ip_report]{virustotal_ip_report()}},
\code{\link[=virustotal_url_scan]{virustotal_url_scan()}}
}
\concept{response classes}
\keyword{internal}
