% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{virustotal_ip_report}
\alias{virustotal_ip_report}
\title{Create a VirusTotal IP report}
\usage{
virustotal_ip_report(data)
}
\arguments{
\item{data}{Raw API response data}
}
\value{
Object of class \code{virustotal_ip_report}
}
\description{
Create a VirusTotal IP report
}
\seealso{
Other response classes:
\code{\link{virustotal-classes}},
\code{\link[=virustotal_domain_report]{virustotal_domain_report()}},
\code{\link[=virustotal_file_report]{virustotal_file_report()}},
\code{\link[=virustotal_file_scan]{virustotal_file_scan()}},
\code{\link[=virustotal_url_scan]{virustotal_url_scan()}}
}
\concept{response classes}
\keyword{internal}
