% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_report.R
\name{domain_report}
\alias{domain_report}
\title{Get Domain Report}
\usage{
domain_report(domain = NULL, ...)
}
\arguments{
\item{domain}{domain name (string)}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
data.frame
}
\description{
Retrieves report on a given domain, including passive DNS, urls detected by at least one url scanner. 
Currently seems to not respond to
}
\examples{
\dontrun{
domain_report("http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

