% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download a file from VirusTotal}
\usage{
download_file(hash = NULL, output_path = NULL, ...)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256)}

\item{output_path}{Local path to save the downloaded file. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
Raw file content or saves file to specified path
}
\description{
Download a file from VirusTotal
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

download_file(hash='99017f6eebbac24f351415dd410d522d',
              output_path='/tmp/downloaded_file')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
