% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_report.R
\name{file_report}
\alias{file_report}
\title{Get File Scan Report}
\usage{
file_report(hash = NULL, ...)
}
\arguments{
\item{hash}{Hash for the scan}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
data.frame with 12 columns: scans, scan_id, sha1, resource, response_code, scan_date
permalink, verbose_msg, total, positives, sha256, md5
}
\description{
Get File Scan Report
}
\examples{
\dontrun{
file_report(hash='99017f6eebbac24f351415dd410d522d')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

