% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_report.R
\name{file_report}
\alias{file_report}
\title{Get File Scan Report}
\usage{
file_report(hash = NULL)
}
\arguments{
\item{hash}{Hash for the scan}
}
\value{
data frame
}
\description{
Get File Scan Report
}
\examples{
\dontrun{
file_report(hash='hash')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

