% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain_votes.R
\name{get_domain_votes}
\alias{get_domain_votes}
\title{Retrieve votes for an Internet domain}
\usage{
get_domain_votes(domain = NULL, limit = NULL, cursor = NULL, ...)
}
\arguments{
\item{domain}{domain name. String. Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
named list
}
\description{
Retrieve votes for an Internet domain
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
get_domain_votes("http://www.google.com")
get_domain_votes("http://www.goodsfwrfw.com") # Domain not found
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
