% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_behaviour_summary.R
\name{get_file_behaviour_summary}
\alias{get_file_behaviour_summary}
\title{Get File Behaviour Summary}
\usage{
get_file_behaviour_summary(hash = NULL, ...)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256). Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing behaviour summary from all sandboxes
}
\description{
Retrieves a summary of all behaviour reports for a file.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_file_behaviour_summary(hash = '99017f6eebbac24f351415dd410d522d')
}
}
\references{
\url{https://docs.virustotal.com/reference/file-all-behaviours-summary}
}
\seealso{
\code{\link{set_key}} for setting the API key,
\code{\link{get_file_behaviours}} for full behaviour reports
}
