% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_behaviours.R
\name{get_file_behaviours}
\alias{get_file_behaviours}
\title{Get File Behaviours}
\usage{
get_file_behaviours(hash = NULL, limit = NULL, cursor = NULL, ...)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256). Required.}

\item{limit}{Number of reports to retrieve. Integer. Optional.}

\item{cursor}{String for pagination. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing behaviour reports from various sandboxes
}
\description{
Retrieves all behaviour reports for a file from various sandboxes.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_file_behaviours(hash = '99017f6eebbac24f351415dd410d522d')
get_file_behaviours(hash = '99017f6eebbac24f351415dd410d522d', limit = 5)
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key,
\code{\link{get_file_behaviour_summary}} for summary,
\code{\link{get_behaviour_report}} for individual sandbox reports
}
