% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_comments.R
\name{get_file_comments}
\alias{get_file_comments}
\title{Retrieve comments for a file}
\usage{
get_file_comments(hash = NULL, limit = NULL, cursor = NULL, ...)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256)}

\item{limit}{Number of comments to retrieve. Integer. Optional. Default is 10.}

\item{cursor}{String for pagination. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing file comments
}
\description{
Retrieve comments for a file
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_file_comments(hash='99017f6eebbac24f351415dd410d522d')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
