% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_download_url.R
\name{get_file_download_url}
\alias{get_file_download_url}
\title{Get download URL for a file}
\usage{
get_file_download_url(hash = NULL, ...)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256)}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing download URL and metadata
}
\description{
Get download URL for a file
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_file_download_url(hash='99017f6eebbac24f351415dd410d522d')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
