% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_relationships.R
\name{get_file_relationships}
\alias{get_file_relationships}
\title{Retrieve relationships for a file}
\usage{
get_file_relationships(
  hash = NULL,
  relationship = NULL,
  limit = NULL,
  cursor = NULL,
  ...
)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256)}

\item{relationship}{Type of relationship: "behaviours", "bundled_files", "compression_parents", "contacted_domains", "contacted_ips", "contacted_urls", "dropped_files", "execution_parents", "itw_domains", "itw_ips", "itw_urls", "overlay_parents", "pcap_parents", "pe_resource_parents", "similar_files", "submissions"}

\item{limit}{Number of relationships to retrieve. Integer. Optional. Default is 10.}

\item{cursor}{String for pagination. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
list containing file relationships
}
\description{
Retrieve relationships for a file
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

get_file_relationships(hash='99017f6eebbac24f351415dd410d522d', 
                       relationship='contacted_domains')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
