% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip_comments.R
\name{get_ip_comments}
\alias{get_ip_comments}
\title{Retrieve comments for an IP address}
\usage{
get_ip_comments(ip = NULL, limit = NULL, ...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}

\item{domain}{domain name. String. Required.}
}
\value{
named list
}
\description{
Retrieves report on a given domain, including passive DNS, urls detected by at least one url scanner. 
Gives category of the domain from bitdefender.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
domains("http://www.google.com")
domains("http://www.goodsfwrfw.com") # Domain not found
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
