% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip_info.R
\name{get_ip_info}
\alias{get_ip_info}
\title{Retrieve information about an IP address}
\usage{
get_ip_info(ip = NULL, limit = NULL, cursor = NULL, ...)
}
\arguments{
\item{ip}{IP address. String. Required.}

\item{limit}{Number of entries. Integer. Optional.  Default is 10.}

\item{cursor}{String. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
named list
}
\description{
Retrieves report on a given domain, including passive DNS, urls detected by at least one url scanner. 
Gives category of the domain from bitdefender.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
get_ip_info("64.233.160.0")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
