% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_report.R, R/ip_report.R
\name{ip_report}
\alias{ip_report}
\title{Get File Scan Report}
\usage{
ip_report(ip = NULL)

ip_report(ip = NULL)
}
\arguments{
\item{ip}{IP Address (String)}

\item{hash}{Hash for the scan}
}
\value{
data frame

data frame
}
\description{
Get File Scan Report

Get URL Report
}
\examples{
\dontrun{
file_report(hash='hash')
}
\dontrun{
ip_report(ip="8.8.8.8")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}

\url{https://www.virustotal.com/en/documentation/public-api/}
}

