% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_report.R
\name{ip_report}
\alias{ip_report}
\title{Get URL Report}
\usage{
ip_report(ip = NULL)
}
\arguments{
\item{ip}{IP Address (String)}
}
\value{
list
}
\description{
Get URL Report
}
\examples{
\dontrun{
ip_report(ip="8.8.8.8")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

