% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limiting.R
\name{is_rate_limit_initialized}
\alias{is_rate_limit_initialized}
\title{Check if rate limiting is properly initialized}
\usage{
is_rate_limit_initialized()
}
\description{
Check if rate limiting is properly initialized
}
\keyword{internal}
