% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_domain_comments.R
\name{post_domain_comments}
\alias{post_domain_comments}
\title{Add a comment to an Internet domain}
\usage{
post_domain_comments(domain = NULL, comment = NULL, ...)
}
\arguments{
\item{domain}{domain name. String. Required.}

\item{comment}{vote. String. Required.  Any word starting with # in your comment's text will be considered a tag, and added to the comment's tag attribute.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
named list
}
\description{
Add a comment to an Internet domain
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
post_domain_comments(domain = "https://google.com", comment = "Great!")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
