% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_domain_votes.R
\name{post_domain_votes}
\alias{post_domain_votes}
\title{Add a vote for a hostname or domain}
\usage{
post_domain_votes(domain = NULL, vote = NULL, ...)
}
\arguments{
\item{domain}{domain name. String. Required.}

\item{vote}{vote. String. Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
named list
}
\description{
Add a vote for a hostname or domain
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
post_domain_votes("http://google.com", vote = "malicious")
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
