% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_file_votes.R
\name{post_file_votes}
\alias{post_file_votes}
\title{Add a vote to a file}
\usage{
post_file_votes(hash = NULL, verdict = NULL, ...)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256)}

\item{verdict}{Vote verdict: "harmless" or "malicious"}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing response data
}
\description{
Add a vote to a file
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

post_file_votes(hash='99017f6eebbac24f351415dd410d522d', verdict='malicious')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
