% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_ip_comments.R
\name{post_ip_comments}
\alias{post_ip_comments}
\title{Add a comment to an IP address}
\usage{
post_ip_comments(ip = NULL, comment = NULL, limit = NULL, ...)
}
\arguments{
\item{ip}{IP address. String. Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}

\item{vote}{vote. String. Required.}
}
\value{
named list
}
\description{
Add a comment to an IP address
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
post_ip_comments(ip = "64.233.160.0", comment = "test")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
