% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_ip_comments.R, R/post_ip_votes.R
\name{post_ip_comments}
\alias{post_ip_comments}
\title{Add a comment to an IP address}
\usage{
post_ip_comments(ip = NULL, comment = NULL, limit = NULL, ...)

post_ip_comments(ip = NULL, comment = NULL, limit = NULL, ...)
}
\arguments{
\item{ip}{IP address. String. Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}

\item{vote}{vote. String. Required.}
}
\value{
named list

named list
}
\description{
Retrieves report on a given domain, including passive DNS, urls detected by at least one url scanner. 
Gives category of the domain from bitdefender.

Retrieves report on a given domain, including passive DNS, urls detected by at least one url scanner. 
Gives category of the domain from bitdefender.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
post_ip_comments("64.233.160.0")
}
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')
   
post_ip_comments("64.233.160.0", comment = "test")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}

\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key

\code{\link{set_key}} for setting the API key
}
