% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{print.virustotal_file_report}
\alias{print.virustotal_file_report}
\title{Print method for file reports}
\usage{
\method{print}{virustotal_file_report}(x, ...)
}
\arguments{
\item{x}{A virustotal_file_report object}

\item{...}{Additional arguments (unused)}
}
\value{
The object, invisibly. Called for the side effect of printing a summary of the file report.
}
\description{
Print method for file reports
}
\keyword{internal}
