% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\name{rate_limit}
\alias{rate_limit}
\title{Rate Limits}
\usage{
rate_limit()
}
\description{
Virustotal requests throttled at 4 per minute. This function creates an env. variable 
that tracks number of requests per minute, and enforces appropriate waiting.
}
