% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescan_file.R
\name{rescan_file}
\alias{rescan_file}
\title{Request rescan of a file}
\usage{
rescan_file(hash = NULL, ...)
}
\arguments{
\item{hash}{File hash (MD5, SHA1, or SHA256) or file ID. String. Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing analysis details and ID
}
\description{
Request a new analysis of a file already present in VirusTotal's database.
Returns an analysis ID that can be used to retrieve the report using \code{\link{file_report}}.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

rescan_file(hash='99017f6eebbac24f351415dd410d522d')
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
