% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescan_file.R
\name{rescan_file}
\alias{rescan_file}
\title{Rescan already submitted files}
\usage{
rescan_file(hash = NULL)
}
\arguments{
\item{hash}{Hash for the scan}
}
\value{
data.frame with 12 columns: scans, scan_id, sha1, resource, response_code, scan_date
permalink, verbose_msg, total, positives, sha256, md5
}
\description{
Rescan already submitted files
}
\examples{
\dontrun{
rescan_file(hash='99017f6eebbac24f351415dd410d522d')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

