% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescan_file.R
\name{rescan_file}
\alias{rescan_file}
\title{Rescan already submitted files}
\usage{
rescan_file(hash = NULL, ...)
}
\arguments{
\item{hash}{Hash for the scan}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
data.frame with 12 columns: 
\code{scans, scan_id, sha1, resource, response_code, scan_date, permalink, verbose_msg, total, positives, sha256, md5}
}
\description{
Rescan already submitted files
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

rescan_file(hash='99017f6eebbac24f351415dd410d522d')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}

