% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescan_ip.R
\name{rescan_ip}
\alias{rescan_ip}
\title{Request rescan of an IP address}
\usage{
rescan_ip(ip = NULL, ...)
}
\arguments{
\item{ip}{IP address to rescan (IPv4 or IPv6). String. Required.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
list containing analysis details and ID
}
\description{
Request a new analysis of an IP address already present in VirusTotal's database.
Returns an analysis ID that can be used to retrieve the report using \code{\link{ip_report}}.
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

# Request rescan of an IPv4 address
rescan_ip("8.8.8.8")

# Request rescan of an IPv6 address
rescan_ip("2001:4860:4860::8888")
}
}
\references{
\url{https://docs.virustotal.com/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key, \code{\link{ip_report}} for getting reports
}
