% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file.R
\name{scan_file}
\alias{scan_file}
\title{Submit file for scanning}
\usage{
scan_file(file_path = NULL, ...)
}
\arguments{
\item{file_path}{Required; Path to the document}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
data.frame with the following columns: scan_id, sha1, resource, response_code, sha256, permalink, md5, verbose_msg
}
\description{
Submit file for scanning
}
\examples{
\dontrun{
scan_file(file_path='path_to_suspicious_file')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

