% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file.R
\name{scan_file}
\alias{scan_file}
\title{Submit a file for scanning}
\usage{
scan_file(file_path = NULL, ...)
}
\arguments{
\item{file_path}{Required; Path to the document}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
data.frame with the following columns: 
\code{scan_id, sha1, resource, response_code, sha256, permalink, md5, verbose_msg}
}
\description{
Submit a file for scanning
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

scan_file(file_path='path_to_suspicious_file')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
\seealso{
\code{\link{set_key}} for setting the API key
}

