% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file.R
\name{scan_file}
\alias{scan_file}
\title{Submit file for scanning}
\usage{
scan_file(file_path = NULL)
}
\arguments{
\item{file_path}{Required; Path to the document}
}
\value{
data.frame
}
\description{
Submit file for scanning
}
\examples{
\dontrun{
scan_file(file_path='path_to_suspicious_file')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

