% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_url.R
\name{scan_url}
\alias{scan_url}
\title{Submit URL for scanning}
\usage{
scan_url(url = NULL, ...)
}
\arguments{
\item{url}{url; string; required}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
data.frame with 7 columns: permalink, resource, url, response_code, scan_date, scan_id, verbose_msg
}
\description{
Submit URL for scanning
}
\examples{
\dontrun{
scan_url("http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

