% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_url.R
\name{scan_url}
\alias{scan_url}
\title{Submit URL for scanning}
\usage{
scan_url(url = NULL, ...)
}
\arguments{
\item{url}{url; string; required}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_POST}}.}
}
\value{
data.frame with 7 columns: 
\code{permalink, resource, url, response_code, scan_date, scan_id, verbose_msg}
}
\description{
Submit a URL for scanning. Returns a data.frame with \code{scan_id} which can be used to 
fetch the report using \code{\link{url_report}}
}
\examples{
\dontrun{

# Before calling the function, set the API key using set_key('api_key_here')

scan_url("http://www.google.com")
}
}
\references{
\url{https://developers.virustotal.com/v2.0/reference}
}
\seealso{
\code{\link{set_key}} for setting the API key
}
