% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_url.R
\name{scan_url}
\alias{scan_url}
\title{Submit URL for Scanning}
\usage{
scan_url(url = NULL)
}
\arguments{
\item{url}{url; string; required}
}
\value{
data frame
}
\description{
Submit URL for Scanning
}
\examples{
\dontrun{
scan_url(url="http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

