% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key.R
\name{set_key}
\alias{set_key}
\title{Set API Key}
\usage{
set_key(api_key = NULL)
}
\arguments{
\item{api_key}{String (API Key)}
}
\description{
Set API Key
}
\examples{
\dontrun{
set_token('api_key_here')
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

