% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key.R
\name{set_key}
\alias{set_key}
\title{Set API Key}
\usage{
set_key(api_key = NULL)
}
\arguments{
\item{api_key}{API key. String. Required.}
}
\description{
Before anything else, get the API key from \url{https://www.virustotal.com/en/}.
Next, use \code{\link{set_key}} to store the API key in an environment variable \code{VirustotalToken}. 
Once you have set the API key, you can use any of the functions.
}
\examples{
\dontrun{

set_key('api_key_here')

}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}
