% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_report.R
\name{url_report}
\alias{url_report}
\title{Get URL Report}
\usage{
url_report(url = NULL, scan = 1, ...)
}
\arguments{
\item{url}{url; string; required}

\item{scan}{numeric; optional; when set to 1, submits url for scan if no existing reports are found; default is 1.}

\item{\dots}{Additional arguments passed to \code{\link{virustotal_GET}}.}
}
\value{
data.frame with 12 columns: scan_id, resource, url, response_code, scan_date, permalink, verbose_msg, positives, total, detected, result, detail
}
\description{
Retrieve a scan report for a given URL
}
\examples{
\dontrun{
url_report("http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

