% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_report.R
\name{url_report}
\alias{url_report}
\title{Get URL Report}
\usage{
url_report(url = NULL, scan = 1)
}
\arguments{
\item{url}{url; string; required}

\item{scan}{numeric; optional; when set to 1, submits url for scan if no existing reports are found}
}
\value{
data frame
}
\description{
Retrieve a scan report on a given URL
}
\examples{
\dontrun{
url_report(url="http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

