% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_report.R
\name{url_report}
\alias{url_report}
\title{Get URL Report}
\usage{
url_report(url = NULL)
}
\arguments{
\item{url}{}
}
\value{
data frame
}
\description{
Get URL Report
}
\examples{
\dontrun{
url_report(url="http://www.google.com")
}
}
\references{
\url{https://www.virustotal.com/en/documentation/public-api/}
}

