% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utilities}
\alias{utilities}
\title{Utility Functions for VirusTotal Package}
\description{
Helper functions and utilities for the VirusTotal package.
}
\seealso{
Other utilities: 
\code{\link{cleanup_temp_files}()},
\code{\link{create_safe_temp_dir}()},
\code{\link{format_file_size}()},
\code{\link{is_safe_environment}()},
\code{\link{validate_vt_response}()},
\code{\link{virustotal_info}()},
\code{\link{virustotal_version}()}
}
\concept{utilities}
