% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R, R/zzz.R
\docType{package}
\name{virustotal-package}
\alias{virustotal}
\alias{virustotal-package}
\title{virustotal: Access Virustotal API}
\description{
Access virustotal API. See \url{https://www.virustotal.com/}.
Details about API calls: \url{https://docs.virustotal.com/reference}.

You will need credentials to use this application.
If you haven't already, get the API Key at \url{https://www.virustotal.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/themains/virustotal}
  \item \url{https://themains.github.io/virustotal/}
  \item Report bugs at \url{https://github.com/themains/virustotal/issues}
}

}
\author{
Gaurav Sood
}
\keyword{internal}
