% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\docType{package}
\name{virustotal-package}
\alias{virustotal-package}
\alias{virustotal}
\title{virustotal: Access Virustotal API}
\description{
Access virustotal API. See \url{https://www.virustotal.com/}. 
Details about results of calls to the API can be found at \url{https://developers.virustotal.com/v2.0/reference}.

You will need credentials to use this application. 
If you haven't already, get the API Key at \url{https://www.virustotal.com/}.
}
\author{
Gaurav Sood
}
