% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\docType{package}
\name{virustotal-package}
\alias{virustotal}
\alias{virustotal-package}
\title{virustotal: Access Virustotal API}
\description{
Get Description of Images. Understand what is in the images. See \url{http://clarifai.com/}. 
Details about results of calls to the API can be found at \url{https://developer.clarifai.com}.

Your need credentials to use this application. 
If you haven't already, you can get this at \url{https://developer.clarifai.com/}.
}
\author{
Gaurav Sood
}
\seealso{
\code{\link{get_token}}
}

