% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{virustotal_auth_error}
\alias{virustotal_auth_error}
\title{Create an authentication error}
\usage{
virustotal_auth_error(
  message = "Invalid or missing API key",
  call = sys.call(-1)
)
}
\arguments{
\item{message}{Error message}

\item{call}{The calling function (automatically detected)}
}
\value{
An error object of class \code{virustotal_auth_error}
}
\description{
Create an authentication error
}
\seealso{
Other error handling: 
\code{\link{virustotal-errors}},
\code{\link{virustotal_check}()},
\code{\link{virustotal_error}()},
\code{\link{virustotal_rate_limit_error}()},
\code{\link{virustotal_validation_error}()}
}
\concept{error handling}
\keyword{internal}
