% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\name{virustotal_check}
\alias{virustotal_check}
\title{Request Response Verification}
\usage{
virustotal_check(req)
}
\arguments{
\item{req}{request}
}
\value{
in case of failure, a message
}
\description{
Request Response Verification
}
