% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{virustotal_file_report}
\alias{virustotal_file_report}
\title{Create a VirusTotal file report}
\usage{
virustotal_file_report(data)
}
\arguments{
\item{data}{Raw API response data}
}
\value{
Object of class \code{virustotal_file_report}
}
\description{
Create a VirusTotal file report
}
\seealso{
Other response classes: 
\code{\link{virustotal-classes}},
\code{\link{virustotal_domain_report}()},
\code{\link{virustotal_file_scan}()},
\code{\link{virustotal_ip_report}()},
\code{\link{virustotal_url_scan}()}
}
\concept{response classes}
\keyword{internal}
