% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\name{virustotal_GET}
\alias{virustotal_GET}
\title{Base POST AND GET functions. Not exported.}
\usage{
virustotal_GET(path, query = list(), key = Sys.getenv("VirustotalToken"), ...)
}
\arguments{
\item{path}{path to the specific API service url}

\item{query}{query list}

\item{key}{A character string containing Virustotal API Key.
Default: \code{Sys.getenv("VirustotalToken")}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
list
}
\description{
GET for the Current V3 API
}
\keyword{internal}
