% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\name{virustotal_POST}
\alias{virustotal_POST}
\title{POST for the Current V3 API}
\usage{
virustotal_POST(
  path,
  body = NULL,
  query = list(),
  key = Sys.getenv("VirustotalToken"),
  encode = "json",
  ...
)
}
\arguments{
\item{path}{path to the specific API service url}

\item{body}{request body (file upload or JSON data)}

\item{query}{query list}

\item{key}{A character string containing Virustotal API Key.
Default: \code{Sys.getenv("VirustotalToken")}.}

\item{encode}{Body encoding, passed to \code{\link[httr]{POST}}. One of
\code{"json"} (the default), \code{"form"} or \code{"multipart"}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
list
}
\description{
POST for the Current V3 API
}
\keyword{internal}
