% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustotal.R
\name{virustotal_POST}
\alias{virustotal_POST}
\title{Base POST AND GET functions. Not exported.}
\usage{
virustotal_POST(query, path = path, key = Sys.getenv("VirustotalToken"),
  ...)
}
\arguments{
\item{query}{query list}

\item{path}{path to the specific API service url}

\item{key}{A character string containing an AWS Access Key ID. The default is retrieved from \code{Sys.getenv("VirustotalToken")}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
list
}
\description{
POST
}

