% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{virustotal_validation_error}
\alias{virustotal_validation_error}
\title{Create a validation error}
\usage{
virustotal_validation_error(
  message,
  parameter = NULL,
  value = NULL,
  call = sys.call(-1)
)
}
\arguments{
\item{message}{Error message}

\item{parameter}{The parameter that failed validation}

\item{value}{The invalid value}

\item{call}{The calling function (automatically detected)}
}
\value{
An error object of class \code{virustotal_validation_error}
}
\description{
Create a validation error
}
\seealso{
Other error handling: 
\code{\link{virustotal-errors}},
\code{\link{virustotal_auth_error}()},
\code{\link{virustotal_check}()},
\code{\link{virustotal_error}()},
\code{\link{virustotal_rate_limit_error}()}
}
\concept{error handling}
